+++ src/os2datascanner/projects/report/reportapp/tests/test_utils.py
@@ -30,6 +30,13 @@ class TestUtils:
         cls.account_sam = account_sam
         cls.user_sam = account_sam.user
 
+        account_jack = Account.objects.create(username="sam_jack",
+                                              first_name="Jack",
+                                              last_name="Samurai",
+                                              organization=org)
+        cls.account_jack = account_jack
+        cls.user_jack = account_jack.user
+
     def test_email_alias_is_created(self, saml_user_data):
         # Arrange / Act
         get_or_create_user_aliases(saml_user_data)
@@ -83,16 +90,16 @@ class TestUtils:
 
     def test_existing_aliases_only_related_to_user(self, saml_user_data):
         # Arrange
-        Alias.objects.create(
+        Alias(
             user=self.user_sam,
             _alias_type=AliasType.SID,
             _value="S-DIG"
-        )
-        Alias.objects.create(
+        ).save(prevent_mismatch=False)  # Needed to force creation of alias without account
+        Alias(
             user=self.user_sam,
             _alias_type=AliasType.EMAIL,
             _value="sam_single@saml.com"
-        )
+        ).save(prevent_mismatch=False)  # Needed to force creation of alias without account
 
         # Act
         get_or_create_user_aliases(saml_user_data)
@@ -129,12 +136,114 @@ class TestUtils:
             _alias_type=AliasType.SID,
             _value="S-DIG"
         )
-        Alias.objects.create(
+        Alias(
             user=self.user_sam,
             _alias_type=AliasType.SID,
             _value="S-DIG"
-        )
+        ).save(prevent_mismatch=False)  # Needed to force creation of alias without account
         # Act
         get_or_create_user_aliases(saml_user_data)
         # Assert
         assert Alias.objects.filter(_alias_type=AliasType.SID).count() == 1
+
+    def test_creating_alias_with_no_account(self):
+        """Creating an alias with no account should throw an exception."""
+        with pytest.raises(IntegrityError):
+            Alias.objects.create(
+                user=self.user_sam,
+                _alias_type=AliasType.SID,
+                _value="S-DIG"
+                )
+
+    def test_updating_aliases_with_a_mismatched_account(self):
+        # Arrange
+        Alias.objects.create(
+            user=self.user_sam,
+            account=self.account_sam,
+            _alias_type=AliasType.SID,
+            _value="S-DIG")
+        Alias.objects.create(
+            user=self.user_sam,
+            account=self.account_sam,
+            _alias_type=AliasType.EMAIL,
+            _value="sam@sam.sam")
+        Alias.objects.create(
+            user=self.user_jack,
+            account=self.account_jack,
+            _alias_type=AliasType.EMAIL,
+            _value="jack@samurai.co.uk")
+
+        with pytest.raises(IntegrityError):  # Assert
+            # Act
+            Alias.objects.all().update(account=self.account_jack)
+
+    def test_bulk_creating_mismatched_aliases(self):
+        # Arrange
+        objs = [
+            Alias(
+                user=self.user_sam,
+                account=self.account_sam,
+                _alias_type=AliasType.SID,
+                _value="S-DIG"),
+            Alias(
+                user=self.user_sam,
+                account=self.account_sam,
+                _alias_type=AliasType.EMAIL,
+                _value="sam@sam.sam"),
+            Alias(
+                user=self.user_jack,
+                account=self.account_jack,
+                _alias_type=AliasType.EMAIL,
+                _value="jack@samurai.co.uk"),
+            Alias(
+                user=self.user_jack,
+                account=self.account_sam,
+                _alias_type=AliasType.EMAIL,
+                _value="jack@sam.sam")]
+
+        with pytest.raises(IntegrityError):  # Assert
+            # Act
+            Alias.objects.bulk_create(objs, ignore_conflicts=False)
+
+    def test_bulk_creating_mismatched_aliases_ignore_conflicts(self):
+        # Arrange
+        objs = [
+            Alias(
+                user=self.user_sam,
+                account=self.account_sam,
+                _alias_type=AliasType.SID,
+                _value="S-DIG"),
+            Alias(
+                user=self.user_sam,
+                account=self.account_sam,
+                _alias_type=AliasType.EMAIL,
+                _value="sam@sam.sam"),
+            Alias(
+                user=self.user_jack,
+                account=self.account_jack,
+                _alias_type=AliasType.EMAIL,
+                _value="jack@samurai.co.uk"),
+            Alias(
+                user=self.user_jack,
+                account=self.account_sam,
+                _alias_type=AliasType.EMAIL,
+                _value="jack@sam.sam")]
+
+        # Act
+        Alias.objects.bulk_create(objs, ignore_conflicts=True)
+
+        # Assert
+        # Only three of the four aliases should be created
+        assert list(Alias.objects.all()) == objs[:3]
+
+    def test_saving_mismatched_alias(self):
+        # Arrange
+        mismatched_alias = Alias(
+            user=self.user_jack,
+            account=self.account_sam,
+            _alias_type=AliasType.EMAIL,
+            _value="jack@sam.sam")
+
+        with pytest.raises(IntegrityError):  # Assert
+            # Act
+            mismatched_alias.save()